Documentation, refactoring and convenience
Moves the annotations into `paralegal-flow`
Provides basic documentation for `paralegal-spdg`
Now requires documentation for `paralegal-spdg` in CI
Adds convenience functions to `Context`
