+++ crates/paralegal-flow/Cargo.toml
@@ -40,6 +40,7 @@ num-derive = "0.4"
 num-traits = "0.2"
 petgraph = { workspace = true }
 humantime = "2"
+strum = { version = "0.25", features = ["derive"] }
 
 
 #dot = "0.1"
